refactor(android): move uploadToServer to Endpoint
